refactor(zainod): move config generation to config module
Relocates GENERATED_CONFIG_HEADER, generate_default_config(), and
related tests from lib.rs to config.rs where they belong.
